Make SuccessDialog global
