+++ lib/screens/upload.dart
@@ -1,6 +1,7 @@
 import 'package:chefio/theme/colors.dart';
 import 'package:chefio/theme/text_styles.dart';
 import 'package:chefio/widgets/buttons.dart';
+import 'package:chefio/widgets/dialogs.dart';
 import 'package:chefio/widgets/input.dart';
 import 'package:chefio/widgets/scaffolds.dart';
 import 'package:flutter/cupertino.dart';
@@ -454,58 +455,15 @@ class _BottomBar extends StatelessWidget {
         onTap: () {
           showCupertinoDialog(
             context: context,
-            builder: (ctx) => const _SuccessDialog(),
+            builder: (ctx) => SuccessDialog(
+              title: 'Upload Success',
+              subtitle:
+                  'Your recipe has been uploaded, you can see it on your profile',
+              buttonText: 'Back to Home',
+              onTap: () => Navigator.pop(context),
+            ),
           );
         },
-      ),
-    );
-  }
-}
-
-class _SuccessDialog extends StatelessWidget {
-  const _SuccessDialog({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      backgroundColor: AppColors.white,
-      child: Padding(
-        padding: EdgeInsets.all(48.r),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(
-              Icons.check_circle_outline_outlined,
-              size: 140.sp,
-              color: AppColors.primary,
-            ),
-            SizedBox(
-              height: 32.h,
-            ),
-            Text(
-              'Upload Success',
-              style: TxtThemes.h1.copyWith(
-                color: AppColors.primaryText.withAlpha(228),
-              ),
-            ),
-            SizedBox(height: 8.h),
-            Text(
-              'Your recipe has been uploaded, you can see it on your profile',
-              textAlign: TextAlign.center,
-              style: TxtThemes.p2.copyWith(color: AppColors.primaryText),
-            ),
-            SizedBox(height: 24.h),
-            BtnPrimary(
-              txt: 'Back to Home',
-              onTap: () {
-                Navigator.pop(context);
-              },
-            )
-          ],
-        ),
-      ),
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(24.r),
       ),
     );
   }
